+++ eslint.config.mjs
@@ -14,7 +14,6 @@ import perfectionistPlugin from 'eslint-plugin-perfectionist';
 import reactPlugin from 'eslint-plugin-react';
 import reactHooksPlugin from 'eslint-plugin-react-hooks';
 import regexpPlugin from 'eslint-plugin-regexp';
-import sonarjsPlugin from 'eslint-plugin-sonarjs';
 import unicornPlugin from 'eslint-plugin-unicorn';
 import globals from 'globals';
 import url from 'node:url';
@@ -43,7 +42,6 @@ export default tseslint.config(
       ['react']: reactPlugin,
       ['react-hooks']: fixupPluginRules(reactHooksPlugin),
       ['regexp']: regexpPlugin,
-      ['sonarjs']: sonarjsPlugin,
       ['unicorn']: unicornPlugin,
     },
     /* eslint-enable no-useless-computed-key */
@@ -190,6 +188,7 @@ export default tseslint.config(
       // Internal repo rules
       //
 
+      '@typescript-eslint/internal/eqeq-nullish': 'error',
       '@typescript-eslint/internal/no-poorly-typed-ts-props': 'error',
       '@typescript-eslint/internal/no-relative-paths-to-internal-packages':
         'error',
@@ -210,7 +209,12 @@ export default tseslint.config(
       ],
       'logical-assignment-operators': 'error',
       'no-console': 'error',
-      'no-else-return': 'error',
+      'no-else-return': [
+        'error',
+        {
+          allowElseIf: false,
+        },
+      ],
       'no-fallthrough': [
         'error',
         { commentPattern: '.*intentional fallthrough.*' },
@@ -335,8 +339,6 @@ export default tseslint.config(
       'regexp/prefer-quantifier': 'error',
       'regexp/prefer-question-quantifier': 'error',
       'regexp/prefer-w': 'error',
-
-      'sonarjs/no-duplicated-branches': 'error',
 
       //
       // eslint-plugin-unicorn