+++ packages/eslint-plugin/src/rules/consistent-type-imports.ts
@@ -669,7 +669,9 @@ export default createRule<Options, MessageIds>({
           yield* fixInsertTypeSpecifierForImportDeclaration(fixer, node, false);
         }
         return;
-      } else if (defaultSpecifier) {
+      }
+
+      if (defaultSpecifier) {
         if (
           report.typeSpecifiers.includes(defaultSpecifier) &&
           namedSpecifiers.length === 0 &&
@@ -678,7 +680,9 @@ export default createRule<Options, MessageIds>({
           // import Type from 'foo'
           yield* fixInsertTypeSpecifierForImportDeclaration(fixer, node, true);
           return;
-        } else if (
+        }
+
+        if (
           fixStyle === 'inline-type-imports' &&
           !report.typeSpecifiers.includes(defaultSpecifier) &&
           namedSpecifiers.length > 0 &&
@@ -699,7 +703,9 @@ export default createRule<Options, MessageIds>({
           // import {AValue, Type1, Type2} from 'foo'
           yield* fixInlineTypeImportDeclaration(fixer, report, sourceImports);
           return;
-        } else if (
+        }
+
+        if (
           namedSpecifiers.every(specifier =>
             report.typeSpecifiers.includes(specifier),
           )