+++ packages/eslint-plugin/src/rules/no-base-to-string.ts
@@ -1,10 +1,16 @@
-/* eslint-disable @typescript-eslint/internal/prefer-ast-types-enum */
 import type { TSESTree } from '@typescript-eslint/utils';
 
 import { AST_NODE_TYPES } from '@typescript-eslint/utils';
+import * as tsutils from 'ts-api-utils';
 import * as ts from 'typescript';
 
-import { createRule, getParserServices, getTypeName } from '../util';
+import {
+  createRule,
+  getConstrainedTypeAtLocation,
+  getParserServices,
+  getTypeName,
+  nullThrows,
+} from '../util';
 
 enum Usefulness {
   Always = 'always',
@@ -12,12 +18,12 @@ enum Usefulness {
   Sometimes = 'may',
 }
 
-type Options = [
+export type Options = [
   {
     ignoredTypeNames?: string[];
   },
 ];
-type MessageIds = 'baseToString';
+export type MessageIds = 'baseArrayJoin' | 'baseToString';
 
 export default createRule<Options, MessageIds>({
   name: 'no-base-to-string',
@@ -30,6 +36,8 @@ export default createRule<Options, MessageIds>({
       requiresTypeChecking: true,
     },
     messages: {
+      baseArrayJoin:
+        "Using `join()` for {{name}} {{certainty}} use Object's default stringification format ('[object Object]') when stringified.",
       baseToString:
         "'{{name}}' {{certainty}} use Object's default stringification format ('[object Object]') when stringified.",
     },
@@ -60,11 +68,10 @@ export default createRule<Options, MessageIds>({
     const checker = services.program.getTypeChecker();
     const ignoredTypeNames = option.ignoredTypeNames ?? [];
 
-    function checkExpression(node: TSESTree.Node, type?: ts.Type): void {
+    function checkExpression(node: TSESTree.Expression, type?: ts.Type): void {
       if (node.type === AST_NODE_TYPES.Literal) {
         return;
       }
-
       const certainty = collectToStringCertainty(
         type ?? services.getTypeAtLocation(node),
       );
@@ -82,16 +89,110 @@ export default createRule<Options, MessageIds>({
       });
     }
 
-    function collectToStringCertainty(type: ts.Type): Usefulness {
-      const toString =
-        checker.getPropertyOfType(type, 'toString') ??
-        checker.getPropertyOfType(type, 'toLocaleString');
-      const declarations = toString?.getDeclarations();
-      if (!toString || !declarations || declarations.length === 0) {
+    function checkExpressionForArrayJoin(
+      node: TSESTree.Node,
+      type: ts.Type,
+    ): void {
+      const certainty = collectJoinCertainty(type);
+
+      if (certainty === Usefulness.Always) {
+        return;
+      }
+
+      context.report({
+        node,
+        messageId: 'baseArrayJoin',
+        data: {
+          name: context.sourceCode.getText(node),
+          certainty,
+        },
+      });
+    }
+
+    function collectUnionTypeCertainty(
+      type: ts.UnionType,
+      collectSubTypeCertainty: (type: ts.Type) => Usefulness,
+    ): Usefulness {
+      const certainties = type.types.map(t => collectSubTypeCertainty(t));
+      if (certainties.every(certainty => certainty === Usefulness.Never)) {
+        return Usefulness.Never;
+      }
+
+      if (certainties.every(certainty => certainty === Usefulness.Always)) {
         return Usefulness.Always;
       }
 
-      // Patch for old version TypeScript, the Boolean type definition missing toString()
+      return Usefulness.Sometimes;
+    }
+
+    function collectIntersectionTypeCertainty(
+      type: ts.IntersectionType,
+      collectSubTypeCertainty: (type: ts.Type) => Usefulness,
+    ): Usefulness {
+      for (const subType of type.types) {
+        const subtypeUsefulness = collectSubTypeCertainty(subType);
+
+        if (subtypeUsefulness === Usefulness.Always) {
+          return Usefulness.Always;
+        }
+      }
+
+      return Usefulness.Never;
+    }
+
+    function collectTupleCertainty(type: ts.TypeReference): Usefulness {
+      const typeArgs = checker.getTypeArguments(type);
+      const certainties = typeArgs.map(t => collectToStringCertainty(t));
+      if (certainties.some(certainty => certainty === Usefulness.Never)) {
+        return Usefulness.Never;
+      }
+
+      if (certainties.some(certainty => certainty === Usefulness.Sometimes)) {
+        return Usefulness.Sometimes;
+      }
+
+      return Usefulness.Always;
+    }
+
+    function collectArrayCertainty(type: ts.Type): Usefulness {
+      const elemType = nullThrows(
+        type.getNumberIndexType(),
+        'array should have number index type',
+      );
+      return collectToStringCertainty(elemType);
+    }
+
+    function collectJoinCertainty(type: ts.Type): Usefulness {
+      if (tsutils.isUnionType(type)) {
+        return collectUnionTypeCertainty(type, collectJoinCertainty);
+      }
+
+      if (tsutils.isIntersectionType(type)) {
+        return collectIntersectionTypeCertainty(type, collectJoinCertainty);
+      }
+
+      if (checker.isTupleType(type)) {
+        return collectTupleCertainty(type);
+      }
+
+      if (checker.isArrayType(type)) {
+        return collectArrayCertainty(type);
+      }
+
+      return Usefulness.Always;
+    }
+
+    function collectToStringCertainty(type: ts.Type): Usefulness {
+      if (tsutils.isTypeParameter(type)) {
+        const constraint = type.getConstraint();
+        if (constraint) {
+          return collectToStringCertainty(constraint);
+        }
+        // unconstrained generic means `unknown`
+        return Usefulness.Always;
+      }
+
+      // the Boolean type definition missing toString()
       if (
         type.flags & ts.TypeFlags.Boolean ||
         type.flags & ts.TypeFlags.BooleanLiteral
@@ -103,64 +204,57 @@ export default createRule<Options, MessageIds>({
         return Usefulness.Always;
       }
 
-      if (
-        declarations.every(
-          ({ parent }) =>
-            !ts.isInterfaceDeclaration(parent) || parent.name.text !== 'Object',
-        )
-      ) {
-        return Usefulness.Always;
-      }
-
       if (type.isIntersection()) {
-        for (const subType of type.types) {
-          const subtypeUsefulness = collectToStringCertainty(subType);
-
-          if (subtypeUsefulness === Usefulness.Always) {
-            return Usefulness.Always;
-          }
-        }
-
-        return Usefulness.Never;
+        return collectIntersectionTypeCertainty(type, collectToStringCertainty);
       }
 
-      if (!type.isUnion()) {
-        return Usefulness.Never;
+      if (type.isUnion()) {
+        return collectUnionTypeCertainty(type, collectToStringCertainty);
       }
 
-      let allSubtypesUseful = true;
-      let someSubtypeUseful = false;
-
-      for (const subType of type.types) {
-        const subtypeUsefulness = collectToStringCertainty(subType);
-
-        if (subtypeUsefulness !== Usefulness.Always && allSubtypesUseful) {
-          allSubtypesUseful = false;
-        }
-
-        if (subtypeUsefulness !== Usefulness.Never && !someSubtypeUseful) {
-          someSubtypeUseful = true;
-        }
+      if (checker.isTupleType(type)) {
+        return collectTupleCertainty(type);
       }
 
-      if (allSubtypesUseful && someSubtypeUseful) {
+      if (checker.isArrayType(type)) {
+        return collectArrayCertainty(type);
+      }
+
+      const toString =
+        checker.getPropertyOfType(type, 'toString') ??
+        checker.getPropertyOfType(type, 'toLocaleString');
+      if (!toString) {
+        // e.g. any/unknown
         return Usefulness.Always;
       }
 
-      if (someSubtypeUseful) {
-        return Usefulness.Sometimes;
+      const declarations = toString.getDeclarations();
+
+      if (declarations == null || declarations.length !== 1) {
+        // If there are multiple declarations, at least one of them must not be
+        // the default object toString.
+        //
+        // This may only matter for older versions of TS
+        // see https://github.com/typescript-eslint/typescript-eslint/issues/8585
+        return Usefulness.Always;
       }
 
-      return Usefulness.Never;
+      const declaration = declarations[0];
+      const isBaseToString =
+        ts.isInterfaceDeclaration(declaration.parent) &&
+        declaration.parent.name.text === 'Object';
+      return isBaseToString ? Usefulness.Never : Usefulness.Always;
     }
 
     function isBuiltInStringCall(node: TSESTree.CallExpression): boolean {
       if (
         node.callee.type === AST_NODE_TYPES.Identifier &&
+        // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
         node.callee.name === 'String' &&
         node.arguments[0]
       ) {
         const scope = context.sourceCode.getScope(node);
+        // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
         const variable = scope.set.get('String');
         return !variable?.defs.length;
       }
@@ -184,9 +278,19 @@ export default createRule<Options, MessageIds>({
         }
       },
       CallExpression(node: TSESTree.CallExpression): void {
-        if (isBuiltInStringCall(node)) {
+        if (
+          isBuiltInStringCall(node) &&
+          node.arguments[0].type !== AST_NODE_TYPES.SpreadElement
+        ) {
           checkExpression(node.arguments[0]);
         }
+      },
+      'CallExpression > MemberExpression.callee > Identifier[name = "join"].property'(
+        node: TSESTree.Expression,
+      ): void {
+        const memberExpr = node.parent as TSESTree.MemberExpression;
+        const type = getConstrainedTypeAtLocation(services, memberExpr.object);
+        checkExpressionForArrayJoin(memberExpr.object, type);
       },
       'CallExpression > MemberExpression.callee > Identifier[name = /^(toLocaleString|toString)$/].property'(
         node: TSESTree.Expression,
@@ -194,6 +298,7 @@ export default createRule<Options, MessageIds>({
         const memberExpr = node.parent as TSESTree.MemberExpression;
         checkExpression(memberExpr.object);
       },
+
       TemplateLiteral(node: TSESTree.TemplateLiteral): void {
         if (node.parent.type === AST_NODE_TYPES.TaggedTemplateExpression) {
           return;