+++ packages/eslint-plugin/src/rules/no-restricted-types.ts
@@ -18,7 +18,6 @@ type Types = Record<
 
 export type Options = [
   {
-    extendDefaults?: boolean;
     types?: Types;
   },
 ];
@@ -146,7 +145,7 @@ export default createRule<Options, MessageIds>({
     ): void {
       const bannedType = bannedTypes.get(name);
 
-      if (bannedType === undefined || bannedType === false) {
+      if (bannedType == null || bannedType === false) {
         return;
       }