+++ packages/eslint-plugin/tests/rules/no-useless-constructor.test.ts
@@ -209,8 +209,6 @@ class A extends Object {
     `,
   ],
   invalid: [
-    /* eslint-disable @typescript-eslint/no-unnecessary-template-expression -- trailing whitespace */
-
     {
       code: `
 class A {
@@ -432,6 +430,5 @@ ${'  '}
         },
       ],
     },
-    /* eslint-enable @typescript-eslint/no-unnecessary-template-expression */
   ],
 });